+++ packages/base-utils/src/index.ts
@@ -1,4 +1,4 @@
 export { bigInt256ToFieldElems, bigintToBEPadded } from "./bits";
-export { zip, range, groupBy, iterChunks, min } from "./functional";
+export { zip, range, groupBy, iterChunks, min, partition } from "./functional";
 export { numberToStringPadded } from "./strings";
 export { assertOrErr } from "./error";