+++ packages/sdk/src/db/notes.ts
@@ -3,108 +3,140 @@ import {
   AssetTrait,
   IncludedNote,
   NoteTrait,
+  IncludedNoteWithNullifier,
+  IncludedNoteCommitment,
+  Note,
 } from "@nocturne-xyz/primitives";
+import { partition, numberToStringPadded } from "@nocturne-xyz/base-utils";
 import * as JSON from "bigint-json-serialization";
 import { KV, KVStore } from "./kvStore";
 
-type NoteAssetKey = string; // Takes form of NOTES_<address>_<id>
-type AllNotes = Map<NoteAssetKey, IncludedNote[]>;
+const NOTES_BY_INDEX_PREFIX = "NOTES_BY_INDEX";
+const NOTES_BY_ASSET_PREFIX = "NOTES_BY_ASSET";
+const NOTES_BY_NULLIFIER_PREFIX = "NOTES_BY_NULLIFIER";
 
-const NOTES_PREFIX = "NOTES";
+// ceil(log10(2^32))
+const MAX_MERKLE_INDEX_DIGITS = 10;
+
+export type AssetKey = string;
+type AllNotes = Map<AssetKey, IncludedNote[]>;
 
 export class NotesDB {
+  // store the following mappings:
+  //  merkleIndexKey => Note | bigint (note if usable, commitment if it's not. A note is usable IFF the user owns it and it hasn't been nullified yet).
+  //  assetKey => merkleIndex[]
+  //  nullifierKey => merkleIndex
+
   public kv: KVStore;
 
   constructor(kv: KVStore) {
     this.kv = kv;
   }
 
-  /**
-   * Format an `IncludedNote` into its corresponding key in the KV store
-   * It produces a key of the form NOTES_<note.asset>_<note.id>_<sha256(note)>
-   *
-   * @param note the note to format
-   * @returns key the corresponding key for the note
-   */
-  static formatNoteKey(note: IncludedNote): string {
-    return (
-      NotesDB.formatNoteAssetKey(note.asset) + "_" + NoteTrait.sha256(note)
-    );
+  static formatIndexKey(merkleIndex: number): string {
+    return `${NOTES_BY_INDEX_PREFIX}-${numberToStringPadded(
+      merkleIndex,
+      MAX_MERKLE_INDEX_DIGITS
+    )}`;
   }
 
-  /**
-   * Format an `Asset` into a key prefix for the all notes in the KV store for that asset
-   * It produces a key of the form NOTES_<asset.address>_<asset.id>.
-   *
-   * @param asset asset for the note
-   * @returns keyPrefix the key prefix for the asset the note is for
-   */
-  static formatNoteAssetKey(asset: Asset): NoteAssetKey {
-    return (
-      NOTES_PREFIX +
-      "_" +
-      asset.assetType +
-      "_" +
-      asset.assetAddr.toUpperCase() +
-      "_" +
-      asset.id
-    );
+  static formatAssetKey(asset: Asset): string {
+    return `${NOTES_BY_ASSET_PREFIX}-${
+      asset.assetType
+    }-${asset.assetAddr.toUpperCase()}-${asset.id.toString()}`;
   }
 
-  /**
-   * Parse a note asset key into an `Asset`. Expects a key of form NOTES_<address>_<id>
-   * into the `Asset` the note is for
-   *
-   * @returns asset a
-   * @throws Error if the key is not of the correct form
-   */
-  static parseAssetFromNoteAssetKey(key: NoteAssetKey): Asset {
-    const arr = key.split("_");
-    if (arr.length !== 4 || arr[0] !== NOTES_PREFIX) {
-      throw Error(`Invalid note asset key: "${key}"`);
-    }
+  static formatNullifierKey(nullifier: bigint): string {
+    return `${NOTES_BY_NULLIFIER_PREFIX}-${nullifier.toString()}`;
+  }
 
+  static parseIndexKey(key: string): number {
+    return parseInt(key.split("-")[1]);
+  }
+
+  static parseAssetKey(key: string): Asset {
+    const [_, assetType, assetAddr, id] = key.split("-");
     return {
-      assetType: AssetTrait.parseAssetType(arr[1]),
-      assetAddr: arr[2],
-      id: BigInt(arr[3]),
+      assetType: AssetTrait.parseAssetType(assetType),
+      assetAddr,
+      id: BigInt(id),
     };
   }
 
-  /**
-   * Store a note
-   *
-   * @param note the note to store
-   * @returns true if successful, false otherwise
-   */
-  async storeNote(note: IncludedNote): Promise<boolean> {
-    const key = NotesDB.formatNoteKey(note);
+  async storeNotesAndCommitments(
+    notesAndCommitments: (IncludedNoteWithNullifier | IncludedNoteCommitment)[]
+  ): Promise<void> {
+    // partition the notes and commitments
+    const [commitmentsPartition, notesPartition] = partition(
+      notesAndCommitments,
+      NoteTrait.isCommitment
+    );
+    const commitments = commitmentsPartition as IncludedNoteCommitment[];
+    const notes = notesPartition as IncludedNoteWithNullifier[];
 
-    if (await this.kv.containsKey(key)) {
-      return true;
-    }
+    // make note and commitment KVs
+    const noteKVs: KV[] = notes.map((note) =>
+      NotesDB.makeNoteKV(note.merkleIndex, note)
+    );
+    const commitmentKVs: KV[] = commitments.map(
+      ({ merkleIndex, noteCommitment }) =>
+        NotesDB.makeCommitmentKV(merkleIndex, noteCommitment)
+    );
 
-    const value = JSON.stringify(note);
-    return await this.kv.putString(key, value);
-  }
+    // make the nullifier => merkleIndex KV pairs
+    const nullifierKVs: KV[] = notes.map(({ merkleIndex, nullifier }) =>
+      NotesDB.makeNullifierKV(merkleIndex, nullifier)
+    );
 
-  /**
-   * Get a note
-   * @param key the key of the note to get
-   * @returns note the note, or undefined if not found
-   * @throws Error if the note is found but is not a valid `IncludedNote`
-   */
-  async removeNote(note: IncludedNote): Promise<boolean> {
-    const key = NotesDB.formatNoteKey(note);
-    return await this.kv.remove(key);
-  }
+    // get the updated asset => merkleIndex[] KV pairs
+    const assetKVs = await this.getUpdatedAssetKVsWithNotesAdded(notes);
 
-  async storeNotes(notes: IncludedNote[]): Promise<boolean> {
-    const kvs: KV[] = notes.map((note) => [
-      NotesDB.formatNoteKey(note),
-      JSON.stringify(note),
+    // write them all into the KV store
+    await this.kv.putMany([
+      ...noteKVs,
+      ...commitmentKVs,
+      ...nullifierKVs,
+      ...assetKVs,
     ]);
-    return await this.kv.putMany(kvs);
+  }
+
+  async nullifyNotes(nullifiers: bigint[]): Promise<void> {
+    // delete nullifier => merkleIndex KV pairs
+    const nfKeys = nullifiers.map((nullifier) =>
+      NotesDB.formatNullifierKey(nullifier)
+    );
+    const kvs = await this.kv.getMany(nfKeys);
+    await this.kv.removeMany([...nfKeys]);
+
+    // get the notes we're nullifying
+    const indices = kvs.map(([_nfKey, stringifiedIdx]) =>
+      parseInt(stringifiedIdx)
+    );
+    const notes = await this.getNotesByMerkleIndices(indices);
+
+    // make the merkleIndex => commitment KV pairs
+    const commitmentKVs: KV[] = notes.map((note) =>
+      NotesDB.makeCommitmentKV(note.merkleIndex, NoteTrait.toCommitment(note))
+    );
+
+    // get the updated asset => merkleIndex[] KV pairs
+    // for each note, remove the note's merkleIndex from its asset's index keys
+    const assetKVs = await this.getUpdatedAssetKVsWithNotesRemoved(notes);
+
+    // write the new commitment KV pairs and the new asset => merkleIndex[] KV pairs to the KV store
+    await this.kv.putMany([...commitmentKVs, ...assetKVs]);
+  }
+
+  /**
+   * Get all notes for an asset
+   *
+   * @param asset the asset to get notes for
+   * @returns notes an array of notes for the asset. The array has no guaranteed order.
+   */
+  async getNotesForAsset(asset: Asset): Promise<IncludedNote[]> {
+    const indices = await this.getMerkleIndicesForAsset(asset);
+
+    return await this.getNotesByMerkleIndices(indices);
   }
 
   /**
@@ -114,38 +146,102 @@ export class NotesDB {
    *          and values are an array of `IncludedNote`s for that asset. The array has no guaranteed order.
    */
   async getAllNotes(): Promise<AllNotes> {
-    const allNotes = new Map<NoteAssetKey, IncludedNote[]>();
+    const allNotes = new Map<AssetKey, IncludedNote[]>();
 
-    const iterPrefix = await this.kv.iterPrefix(NOTES_PREFIX);
-    for await (const [, value] of iterPrefix) {
-      const note = JSON.parse(value);
-      const noteAssetKey = NotesDB.formatNoteAssetKey(note.asset);
+    const iterPrefix = await this.kv.iterPrefix(NOTES_BY_ASSET_PREFIX);
+    for await (const [assetKey, stringifiedIndices] of iterPrefix) {
+      const indices: number[] = JSON.parse(stringifiedIndices);
+      const notes = await this.getNotesByMerkleIndices(indices);
 
-      const notesForAsset = allNotes.get(noteAssetKey) ?? [];
-      notesForAsset.push(note);
+      const notesForAsset = allNotes.get(assetKey) ?? [];
+      notesForAsset.push(...notes);
 
-      allNotes.set(noteAssetKey, notesForAsset);
+      if (notesForAsset.length !== 0) {
+        allNotes.set(assetKey, notesForAsset);
+      }
     }
 
     return allNotes;
   }
 
-  /**
-   * Get all notes for an asset
-   *
-   * @param asset the asset to get notes for
-   * @returns notes an array of notes for the asset. The array has no guaranteed order.
-   */
-  async getNotesFor(asset: Asset): Promise<IncludedNote[]> {
-    const noteAssetKey = NotesDB.formatNoteAssetKey(asset);
+  private static makeCommitmentKV(merkleIndex: number, commitment: bigint): KV {
+    return [NotesDB.formatIndexKey(merkleIndex), commitment.toString()];
+  }
 
-    const notes = [];
-    const iterPrefix = await this.kv.iterPrefix(noteAssetKey);
-    for await (const [, value] of iterPrefix) {
-      const note = JSON.parse(value);
-      notes.push(note);
+  private static makeNoteKV<N extends Note>(merkleIndex: number, note: N): KV {
+    return [
+      NotesDB.formatIndexKey(merkleIndex),
+      JSON.stringify(NoteTrait.toNote(note)),
+    ];
+  }
+
+  private static makeNullifierKV(merkleIndex: number, nullifier: bigint): KV {
+    return [NotesDB.formatNullifierKey(nullifier), merkleIndex.toString()];
+  }
+
+  private async getUpdatedAssetKVsWithNotesAdded<N extends IncludedNote>(
+    notes: N[]
+  ): Promise<KV[]> {
+    const map = new Map<AssetKey, Set<number>>();
+    for (const note of notes) {
+      const assetKey = NotesDB.formatAssetKey(note.asset);
+      let indices = map.get(assetKey);
+      if (!indices) {
+        indices = new Set(await this.getMerkleIndicesForAsset(note.asset));
+      }
+
+      indices.add(note.merkleIndex);
+
+      map.set(assetKey, indices);
     }
 
-    return notes;
+    return Array.from(map.entries()).map(([assetKey, indexKeys]) => [
+      assetKey,
+      JSON.stringify(Array.from(indexKeys)),
+    ]);
+  }
+
+  private async getUpdatedAssetKVsWithNotesRemoved<N extends IncludedNote>(
+    notes: N[]
+  ): Promise<KV[]> {
+    const map = new Map<AssetKey, Set<number>>();
+    for (const note of notes) {
+      const assetKey = NotesDB.formatAssetKey(note.asset);
+      let indices = map.get(assetKey);
+      if (!indices) {
+        indices = new Set(await this.getMerkleIndicesForAsset(note.asset));
+      }
+
+      indices.delete(note.merkleIndex);
+
+      map.set(assetKey, indices);
+    }
+
+    return Array.from(map.entries()).map(([assetKey, indexKeys]) => [
+      assetKey,
+      JSON.stringify(Array.from(indexKeys)),
+    ]);
+  }
+
+  private async getMerkleIndicesForAsset(asset: Asset): Promise<number[]> {
+    const assetKey = NotesDB.formatAssetKey(asset);
+    const value = await this.kv.getString(assetKey);
+    if (!value) {
+      return [];
+    }
+
+    return JSON.parse(value);
+  }
+
+  private async getNotesByMerkleIndices(
+    indices: number[]
+  ): Promise<IncludedNote[]> {
+    const idxKeys = indices.map((index) => NotesDB.formatIndexKey(index));
+    const kvs = await this.kv.getMany(idxKeys);
+    return kvs.map(([key, value]) => {
+      const merkleIndex = NotesDB.parseIndexKey(key);
+      const note = JSON.parse(value);
+      return NoteTrait.toIncludedNote(note, merkleIndex);
+    });
   }
 }